rename: rename frontend service
